[ARM] goldfish: Enable low-level kernel debugging.
This provides a working implementation of the assembly routines required
to have CONFIG_DEBUG_LL working with the goldfish kernel. For the record,
this instructs the kernel to send each printk() to the serial port as soon
as possible, instead of buffering the output.

This turned out valuable to debug some tricky ARMv7 emulation support issues.

This also cleans up a bit existing code related to the goldfish TTY.
Signed-off-by: David 'Digit' Turner <digit@android.com>