mips: Outline udelay and fix a few issues.
Outlining fixes the issue were on certain CPUs such as the R10000 family
the delay loop would need an extra cycle if it overlaps a cacheline
boundary.

The rewrite also fixes build errors with GCC 4.4 which was changed in
way incompatible with the kernel's inline assembly.

Relying on pure C for computation of the delay value removes the need for
explicit.  The price we pay is a slight slowdown of the computation - to
be fixed on another day.

Signed-off-by: Ralf Baechle <ralf@linux-mips.org>

Backported from head of tree for gcc-4.4 compatibility
Additional bugfixes:
MIPS: Fix __ndelay build error and add 'ull' suffix for 32-bit kernel
MIPS: Fix typo resulting in far too long ndelay times.

Change-Id: If6e1ed24bd7dca74425b77b41ceac4cf9abb1941
Signed-off-by: Chris Dearman <chris@mips.com>
Signed-off-by: Raghu Gandham <raghu@mips.com>