goldfish: Fix sd card init failure.
After porting the sd driver from 2.6 to 3.4, during device init an
interrupt would occur with a non-null data pointer and end-of-xfer
set to false. Thereafter, no further interrupts would occur, the
thread driving the initialization would get stuck, etc.

This is the only case I've seen where the data pointer is non-null
and the end-of-xfer flag is false. This change causes the interrupt
handler to behave in that case, and only for a specific op code, as
if the end-of-xfer flag were true.

The problem happens as stack is actually sending ACMD13 while emulator
only supports CMD13. The workaround works but needs better fix in future.

Change-Id: Ifcb0797aad1a2423c972e7aaf8d955a00e14fec7
Signed-off-by: Tom Keel <thomas.keel@intel.com>
Signed-off-by: keunyoung <keunyoung@google.com>