video: goldfishfb: support framebuffer backed by memblock
Compared to the consistent DMA region, memblocks can reserve much larger
areas of memory, which are needed to back high-resolution displays.

If the platform_device has a second memory resource, it refers to a
memblock that goldfishfb can use to back the display (and release any
unused parts).  Otherwise fall back on allocating from the consistent
DMA region.

Change-Id: I6e857b87423f1a7ce51e9addc677341f0251fcfe
Signed-off-by: Greg Hackmann <ghackmann@google.com>